refactor: Update component imports and remove unused files refactor: Update component imports and remove unused files
- Updated import paths for Home and Settings components to reflect new directory structure.
- Removed obsolete Home and Setting component files to streamline the codebase.
- Simplified user data update functions by removing unnecessary response handling.
