+++ packages/Frontend/src/component/pages/Home/Home.jsx
@@ -1,0 +1,45 @@
+import { useState, useEffect } from "react";
+import { motion } from "framer-motion";
+import { Button } from "@mui/material";
+import {
+  CheckCircle,
+  Clock,
+  ListTodo,
+  PlusCircle,
+  Star,
+  TrendingUp,
+} from "lucide-react";
+import { useNavigate } from "react-router-dom";
+import Header from "./components/Header";
+import Hero from "./components/Hero";
+import Stats from "./components/Stats";
+import Features from "./components/Features";
+import Testimonials from "./components/Testimonials";
+import CTA from "./components/CTA";
+import Footer from "./components/Footer";
+
+export default function Home() {
+  const [scrollY, setScrollY] = useState(0);
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setScrollY(window.scrollY);
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, []);
+
+  return (
+    <div className="min-h-screen bg-gradient-to-b from-blue-50 to-white">
+      <Header />
+      <Hero />
+      <Stats />
+      <Features />
+      <Testimonials />
+      <CTA />
+      <Footer />
+    </div>
+  );
+}
