+++ packages/Frontend/src/component/pages/Settings/taps/NotificationSettings.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+const NotificationSettings = () => {
+  return (
+    <div className="bg-white dark:bg-[#080808] dark:text-[#e0e0e0] rounded-lg shadow p-6">
+      <h3 className="text-lg font-medium text-gray-800 dark:text-gray-200 mb-2">
+        Notification Settings
+      </h3>
+      <p className="text-sm text-gray-600 dark:text-gray-300">
+        Configure your notification preferences here.
+      </p>
+    </div>
+  );
+};
+
+export default NotificationSettings;
