--- packages/Frontend/src/stores/projectStore.js
@@ -1,5 +1,0 @@
-import { create } from "zustand";
-
-export const useProjecthStore = create((set) => ({
-
-}));
